First-Deplyment Testing (#4)
* manage-teams-updated

add employee, manage teams updates to align with new rules and stuff

* My Profile Page

My profile page UI and functionality, add updated setup page to insert store id and role id to the database

* dashboard UI

* more features

* Update Dashboard UI

* Update LoginForm.jsx

* improvement on dashboard and manager role added

* fixing home-page

* a little change

* Introduction of Request HandlerReq

- created a request handler to handle employee requests for time-off, complaint and availability change.
- all the request are inserted into the employee request table and will be reviewed by the manager or owner for approval.

* Update Dashboard.jsx

* Update Dashboard.jsx

* time-off requests

- made changes to time-off page to fetch all the request from the employee_request table where the request-type is time-off.
- updated the requestHandler to accomodate the new request type.

* Update Dashboard.jsx

* Added Profile Photo Feature

Added Profile Photo Feature Using Storage Function in Supabase

* Major Bug fixes

I just fixed the email link bug
Bug description - "The link we are creating right now contains the raw details within the link which can be manuplated to change the entry in the employee table as well as the auth table which is a really big vulnerability in the system."

How I fixed it- Created a table to handle the token generated in the link and set it as PK of the table and stored all the values in the database like employee id store id and email now when someone click on the link (that only contains a random generated token) the setup page checks the existence of the token in the database and fetches the related data from there. hence making it more safe and secure

Other fixes - Small UI changes on login screen and the view employee screen

* fixed-the mange stores feature

* Added Automated Mail Send Feature

Used Mailjet API to implement automated Mail Sending Feature

* Add .gitignore and remove node_modules from repo

* fixed-the-backend-folder-not-working -on other-devices

* blank-files

there were some blank files comitted to the branch that led to a huge issue no idea how it happened but now its fixed.

* Update to homepage

* wrong-table-entry in dashboard.jsx

* implemented schedule planner for manager

* timeoff and fetchschedule

* employee schedule

So in this commit I have completed the main page of the employee schedule and most of the availability section is also done and I have done some changes on the employee dashboard as well and on the navigation bar as well

* Time-Off Handling

-Now approve and decline functionalities are working for time-offs and availability changes
- fixed the edit feature in schedule planner and now it updating the old records rather than creating new inserts.
-requestHandler updated to align with approve and decline fuctionalities
-UI changes for request viewer

* final-merger

* final-merger-2

---------

Co-authored-by: Princgrewal <164895925+Princgrewal@users.noreply.github.com>
Co-authored-by: Noor <159936718+noor741@users.noreply.github.com>
Co-authored-by: moin-shk <moin.shaikh2684@gmail.com>