Init pull (#6)
* Updated Dashboard and Employee list

Redesigned the whole UI for the Dashboards of Admin/owners , Manager and all associates, added the search functionality to the employee list and Redesigned the UI for the Navbar - made it vertical instead of horizontal

* resolved

* Added my Store for manager

* updated working code but some UI is old

* updated working code but some UI is old

* calendar-component

created a calendar component for to be used across different pages. fixed the schedule planner UI a lil but still needs alot of updates

* merged

* functional scheduling page

New feature filter added to view individual employee schedule

* clock-in and clock-out functionality

worked on the major clock in and out function the UI for the pages used nominatim API for assigning stores their coordinates and used geolocation api built in library for fetching users live location while using clock in and out functionality

* Update main.jsx

* Updated Auth.Js

Fixed the logout bug
Now the user will only be logged out after 30 mins of the session created when he does not check the keep me logged in check box and incase if he does than 60 days before he is logged out.

---------

Co-authored-by: Noor <159936718+noor741@users.noreply.github.com>
Co-authored-by: Princgrewal <164895925+Princgrewal@users.noreply.github.com>